Add files via upload
This file contains an interface of a mobile application. This application can be used to exchange items with anyone around the globe provided that they are valued at the same price. The price will be determined by a AI algorithm which takes several things into account such as date of manufacturing, Authenticity of the product and Quality of it among other things. People can also put their items on auction so they can potentially profit from a transaction.